[FEATURE] Added combiner class for performance

+++ src/net/moc/LogAnalyzer/BrowserMapper.java
@@ -12,17 +12,17 @@ import nl.bitwalker.useragentutils.*;
 
 
 
-public class BrowserMapper extends Mapper<LongWritable, Text, Text, IntWritable> {
+public class BrowserMapper extends Mapper<LongWritable, Text, Text, LongWritable> {
 	
 	public void map(LongWritable key, Text value, Context context) throws IOException, InterruptedException {
 		try {
 				//@todo: Maybe we should try to see if this is really JSON encoded string or not. Perhaps some sort of factory
 			ApacheCombinedLogEntryInterface logEntry = ApacheCombinedLogEntryFactory.getLogEntryFromString(value.toString());
-			context.write(new Text(logEntry.getUserAgent().getBrowser().getName()) , new IntWritable(1));
+			context.write(new Text(logEntry.getUserAgent().getBrowser().getName()) , new LongWritable(1));
 			
 		} catch ( Exception e) {
 			System.err.println("Error parsing logstash outpu");
-			context.write(new Text("Unknown (parse error in logentry") , new IntWritable(1));
+			context.write(new Text("Unknown (parse error in logentry") , new LongWritable(1));
 		}
 	}
